Add entity keys abstraction
Added a new QupEntity abstract class that uses the reflection to return
a list of all property with the `[Key]` attribute for the required
 `GetKeys` method.

Extracted entity model validation from the yaml repository into an
extension methods that allows the domain model to perform entity model
validation when required.
