+++ src/FuseDigital.QuickSetup.Domain/Projects/Project.cs
@@ -1,20 +1,15 @@
 ﻿using System.ComponentModel.DataAnnotations;
+using FuseDigital.QuickSetup.Entities;
 using FuseDigital.QuickSetup.VersionControlSystems;
-using Volo.Abp.Domain.Entities;
 
 namespace FuseDigital.QuickSetup.Projects;
 
-public class Project : Entity
+public class Project : QupEntity
 {
     [Key]
     public string RelativePath { get; set; }
 
     public string Repository { get; set; }
-
-    public override object[] GetKeys()
-    {
-        return new object[] {RelativePath};
-    }
 
     public void Clone(IVersionControlDomainService versionControl)
     {
